Refeactor file structure
