--- frontend/src/App.js
@@ -1,31 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import LeaderBoard from "./components/LeaderBoard";
-import React from "react";
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <p>
-
-          <LeaderBoard n = {5} />
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
