+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Router from "./Router";
-import './styles.css';
+import Router from "./components/Router.tsx";
+import './styles/style.css';
 
 ReactDOM.createRoot(document.getElementById("root")).render(
     <React.StrictMode>
